+++ src/components/form/BulmaSelect.vue
@@ -1,12 +1,16 @@
 <template>
 	<span :class="classes" class="select">
-<!--		todo select multiple does not update to array-->
 		<select :multiple="!!isMultiple" :size="multipleSize" :value="modelValue" v-bind="$attrs"
 		        @input="transformSelected($event)">
 			<slot>
-				<option v-for="option in options" :value="optionValue(option)">
-					{{ (option.text ? option.text : option) }}
-				</option>
+				<template v-for="item in options">
+					<option v-if="typeof item === 'string'" :value="item">{{ item }}</option>
+					<optgroup v-else-if="isBulmaOptionGroup(item)" :disabled="item.disabled" :label="item.label">
+							<option v-for="option in item.options" :value="option.value ? option.value :option"
+							>{{ option.label ? option.label : option }}</option>
+					</optgroup>
+					<option v-else :disabled="item.disabled" :value="item.value">{{ item.label }}</option>
+				</template>
 			</slot>
 		</select>
 	</span>
@@ -15,6 +19,7 @@
 <script lang="ts" setup>
 	import {computed} from "vue";
 	import {BulmaColour, BulmaInputState, BulmaSize, getBulmaClassesFromProps} from "../../types";
+	import {BulmaOption, BulmaOptionGroup, isBulmaOptionGroup} from "../../types/SelectTypes";
 
 	const props = defineProps<{
 		/** The colour of the select. */
@@ -30,29 +35,23 @@
 		//todo: make a type for this
 		//todo: make type for adding optgroups
 		/** The options to display in the select. v-model only works with options prop */
-		options?: ({ text: string, value: any } | string)[]
-
+		options?: (BulmaOption | BulmaOptionGroup | string)[]
+		/** The initial value of the select. Used for v-model */
 		modelValue?: any
 		/** Must pair with BulmaFormControl with isExpanded prop top work; makes select expand to fill its parent */
-		isFullWidth?: boolean
+		isFullwidth?: boolean
 	}>();
 
 	const emit = defineEmits<{
-		(name: 'update:modelValue', newValue: any | any[]): any
+		(name: 'update:modelValue', newValue: any | any[]): unknown
 	}>();
 	const multipleSize = computed(() => typeof props.isMultiple === 'number' ? props.isMultiple : null);
 	const classes = computed(() => getBulmaClassesFromProps(props));
 
-
-	function optionValue(option: { text: string, value: any } | string) {
-		if(typeof option === 'string') {
-			return option;
-		}
-	}
-
-	//todo: return value always as array, or not idk
 	/**
 	 * Get the array of selected values from the select element
+	 * If the select is not multiple, return the value of the selected option
+	 * else return an array of the values of the selected options
 	 * @param {Event} e - The input event
 	 */
 	function transformSelected(e: Event) {